fix: remove size constant from constant.js and move to breakpoints.js
