+++ src/utils/breakpoints.js
@@ -29,4 +29,13 @@ const breakpoints = {
   },
 };
 
+export const Size = Object.freeze({
+  xs: 'extraSmall',
+  sm: 'small',
+  md: 'medium',
+  lg: 'large',
+  xl: 'extraLarge',
+  xxl: 'extraExtraLarge',
+});
+
 export default breakpoints;
