+++ src/utils/constants.js
@@ -7,13 +7,4 @@ const Variant = Object.freeze({
   },
 });
 
-export const Size = Object.freeze({
-  xs: 'extraSmall',
-  sm: 'small',
-  md: 'medium',
-  lg: 'large',
-  xl: 'extraLarge',
-  xxl: 'extraExtraLarge',
-});
-
 export default Variant;
